scripts/requirements.txt
# Azure Blob Storage SDK
azure-storage-blob==12.19.0

# Azure Identity (for advanced authentication)
azure-identity==1.15.0

# Other utilities
python-dotenv==1.0.0
